update sécurity management extoplasma

+++ app/Controller/UserController.php
@@ -24,15 +24,14 @@ class UserController extends MasterController
 			//On redirige vers l'accueil des utilisateurs connectés
 			$this->redirectToRoute('default_home');
 
-		}else{
+		}
+		else{
 			//instancie variables
 			$post = [];
 			$errors = [];
 			$success = false;
 			$successimg = false;
 			$adress = ''; //adress est visible pour toute la fonction
-
-
 
 			//A l'insertion de l'image dans le formulaire, il est placé dans le fichier assets
 			$folder = $_SERVER['DOCUMENT_ROOT'].'/limonade/public/assets/image/';
@@ -42,11 +41,11 @@ class UserController extends MasterController
 			if(!empty($_FILES) && isset($_FILES['avatar'])){
 
 				// Récupère le nom de mon fichier
-				$nomFichier = $_FILES['avatar']['name']; 
+				$nomFichier = $_FILES['avatar']['name'];
 				// Stockage temporaire du fichier
 				$tmpFichier = $_FILES['avatar']['tmp_name'];
 				// Créer une chaine de caractère contenant le nom du dossier de destination et le nom du fichier final
-				$newFichier = $folder.$nomFichier; 
+				$newFichier = $folder.$nomFichier;
 				// Permet de vérifier que la taille du fichier est inférieure ou égale à $maxSize
 				if($_FILES['avatar']['size'] <= $maxSize){
 					/*
@@ -134,6 +133,8 @@ class UserController extends MasterController
 						$userInserted = $usersModel->getUserByUsernameOrEmail($dataUser['email']);
 						echo $userInserted['id'];
 
+						// we compose a link to send
+						$magicLink = '<a href="http://localhost/limonade/public/registerConfirm?email='.$dataToken['email'].'&token='.$token.'">ici!</a>';
 						$success =  true;
 						$mail = new PHPMailer;
 						//$mail->SMTPDebug = 3;                               // Enable verbose debug output
@@ -152,14 +153,14 @@ class UserController extends MasterController
 						$mail->isHTML(true);                                  // Set email format to HTML
 
 						$mail->Subject = 'Valider votre compte';
-						$mail->Body    =  $dataUser['username'].' Afin de valider votre compte merci de cliquer sur ce lien http://localhost/limonade/public/registerConfirm?email='.$dataToken['email'].'&token='.$token;
-						$mail->AltBody = $dataUser['username'].' Afin de valider votre compte merci de cliquer sur ce lien http://localhost/limonade/public/registerConfirm?email='.$dataToken['email'].'&token='.$token;
+						$mail->Body    =  $dataUser['username'].' Afin de valider votre compte merci de cliquer '.$magicLink;
+						$mail->AltBody = $dataUser['username'].' Afin de valider votre compte merci de cliquer '.$magicLink;
 
 						if(!$mail->send()) {
 								echo 'Message could not be sent.';
 								echo 'Mailer Error: ' . $mail->ErrorInfo;
 						} else {
-								echo 'Message has been sent';
+								// echo 'Message has been sent';
 						}
 
 						//redirige l'utilisateur vers la page d'accueil
@@ -172,7 +173,6 @@ class UserController extends MasterController
 
 			}
 			# On envoi les erreurs en paramètre à l'aide d'un tableau (array)
-
 			$params = ['errors' => $errors, 'success' => $success, 'successimg' => $successimg, 'adress' => $adress];
 			$this->show('user/register', $params);
 		}
@@ -182,59 +182,64 @@ class UserController extends MasterController
 	/**
 	 * Function Pour confirmer le compte
 	 */
-	public function registerConfirm(){
+	public function registerConfirm()
+	{
+		if(!isset($_GET['email']) && !isset($_GET['token'])){
+			$this->redirectToRoute('default_home');
+		}
+		else{
+			$error = [];
+			$validation = false;
 
-		$error = [];
-		$validation = false;
+			if(isset($_GET['token']) &&
+			!empty($_GET['token']) &&
+			isset($_GET['email']) &&
+			!empty($_GET['email'])){
 
-		if(isset($_GET['token']) &&
-		!empty($_GET['token']) &&
-		isset($_GET['email']) &&
-		!empty($_GET['email'])){
+			//On récupère et nettoye les valeur des GET
+			foreach ($_GET as $key => $value) {
+	          	$get[$key] = trim(strip_tags($value));
+	      	}
 
-		//On récupère et nettoye les valeur des GET
-		foreach ($_GET as $key => $value) {
-          	$get[$key] = trim(strip_tags($value));
-      	}
+	      	//On récupère l'email et le token dans une varialbe
+	      	$email = $get['email'];
+	     	$token = $get['token'];
 
-      	//On récupère l'email et le token dans une varialbe
-      	$email = $get['email'];
-     	$token = $get['token'];
+		    //On instancie la class tokensPasswordModel
+			$tokensRegisterModel = new TokensRegisterModel();
 
-	    //On instancie la class tokensPasswordModel
-		$tokensRegisterModel = new TokensRegisterModel();
+		    //On vérifie que le token et l'email sont bien dans la bdd
+	    	$tokenExist = $tokensRegisterModel->findTokenRegister($email, $token);
 
-	    //On vérifie que le token et l'email sont bien dans la bdd
-    	$tokenExist = $tokensRegisterModel->findTokenRegister($email, $token);
+			//Si le token est valide et n'a pas expiré, on peut donc activer le compte
+			if(!empty($tokenExist) && ($tokenExist['date_exp'] > date('Y-m-d H:i:s'))) {
 
-		//Si le token est valide et n'a pas expiré, on peut donc activer le compte
-		if(!empty($tokenExist) && ($tokenExist['date_exp'] > date('Y-m-d H:i:s'))) {
-		
-			// On instancie la classe UsersModel pour utiliser les fonctions du UserSModel
-			$usersModel = new UsersModel();
+				// On instancie la classe UsersModel pour utiliser les fonctions du UserSModel
+				$usersModel = new UsersModel();
 
-			// On récupère : infos de table user, via son adresse mail, cela change la valeur du champ activation de false a true 
-			$infoUser = $usersModel->getUserByUsernameOrEmail($get['email']);
-				$data = [
-					'activation' => 'true'
-				];
+				// On récupère : infos de table user, via son adresse mail, cela change la valeur du champ activation de false a true
+				$infoUser = $usersModel->getUserByUsernameOrEmail($get['email']);
+					$data = [
+						'activation' => 'true'
+					];
 
-			// Le compte est enfin actif
-			if($usersModel->update($data, $infoUser['id'])){
-				// Comme le compte est activé, supression du token
-				if($tokensRegisterModel->delete($tokenExist['id'])){
+				// Le compte est enfin actif
+				if($usersModel->update($data, $infoUser['id'])){
+					// Comme le compte est activé, supression du token
+					if($tokensRegisterModel->delete($tokenExist['id'])){
 
-					echo 'Tout est ok : Supression du token et activation du compte';
+						 echo 'Tout est ok : Supression du token et activation du compte';
+					}
 				}
 			}
-		}
 
-		if($tokenExist['date_exp'] < date('Y-m-d H:i:s')){
-			$error[] = 'Le lien n\'est plus valide ou votre compte est déja activé.' ;
-		}
+			if($tokenExist['date_exp'] < date('Y-m-d H:i:s')){
+				$error[] = 'Le lien n\'est plus valide ou votre compte est déja activé.' ;
+			}
 
+			}
+			$this->show('user/registerConfirm');
 		}
-		$this->show('user/registerConfirm');
 	}
 
 	/**
@@ -249,30 +254,32 @@ class UserController extends MasterController
 			//On redirige vers l'accueil des utilisateurs connectés
 			$this->redirectToRoute('default_home');
 		}else{
-		//on s'identifie vers l'api facebook et on est redirigé vers loginhelper
-		$fb = new \Facebook\Facebook([
-		'app_id' => '602369446588975',
-		'app_secret' => 'dd68c176483e918625b46de16c01419e',
-		'default_graph_version' => 'v2.2',
-		]);
+			//on s'identifie vers l'api facebook et on est redirigé vers loginhelper
+			$fb = new \Facebook\Facebook([
+			'app_id' => '602369446588975',
+			'app_secret' => 'dd68c176483e918625b46de16c01419e',
+			'default_graph_version' => 'v2.2',
+			]);
 
-		$helper = $fb->getRedirectLoginHelper();
+			$helper = $fb->getRedirectLoginHelper();
 
-		// Build URL for callback
-		$urlRedirect = 'http://'.$_SERVER['HTTP_HOST'].$this->generateUrl('user_fbCallBack');
-		$permissions = ['email']; // Optional permissions
+			// Build URL for callback
+			$urlRedirect = 'http://'.$_SERVER['HTTP_HOST'].$this->generateUrl('user_fbCallBack');
+			$permissions = ['email']; // Optional permissions
 
 
-		$urlFacebook = $helper->getLoginUrl($urlRedirect, $permissions);
+			$urlFacebook = $helper->getLoginUrl($urlRedirect, $permissions);
 
-		$this->redirect($urlFacebook);
+			$this->redirect($urlFacebook);
+		}
 	}
-	}
+
+
 	/**
 	 * Fonction Permettant de revenir depuis fb au site web
 	 */
-	public function fbCallBack(){
-
+	public function fbCallBack()
+	{
 		$data = [];
 		$success = false;
 
@@ -366,7 +373,7 @@ class UserController extends MasterController
 	    $passwordHash = $authModel->hashPassword($passwordPlain);
 
 	    if(!empty($user)){
-	    	$data = [ 
+	    	$data = [
 	    		'username' 		=> $user['name'],
 	    		'firstname' 	=> $user['first_name'],
 	    		'lastname' 		=> $user['last_name'],
@@ -407,7 +414,7 @@ class UserController extends MasterController
 					echo 'Message could not be sent.';
 					echo 'Mailer Error: ' . $mail->ErrorInfo;
 			} else {
-					echo 'Message has been sent';
+					//echo 'Message has been sent';
 			}
 
 
@@ -422,13 +429,13 @@ class UserController extends MasterController
 	    // User is logged in with a long-lived access token.
 	    // You can redirect them to a members-only page.
 	    //header('Location: https://example.com/members.php');
-		}
+	}
 
 		/**
 		 * Fonction Permettant de se connecter
 		 */
-	public function login(){
-
+	public function login()
+	{
 		// Si il est conecter on le redirige
 		$loggedUser = $this->getUser();
 
@@ -436,13 +443,11 @@ class UserController extends MasterController
 			//connecté
 		$this->redirectToRoute('default_home');
 		}
-		else
-		{
+		else{
 			$post = [];
 			$errors = [];
 
 			if(!empty($_POST)){
-
 				// On nettoie les données...c'est l'équivalent de notre foreach
 				$post = array_map('strip_tags', $_POST);
 				$post = array_map('trim', $post);
@@ -471,10 +476,12 @@ class UserController extends MasterController
 							// $myUser = $authModel->getLoggedUser(); // Permet de récupérer les infos de sessions
 							// $myUser = $this->getUser(); // Permet de récupérer les infos de sessions
 							$this->redirectToRoute('default_home');
-						}else {
+						}
+						else{
 							$errors[] = 'Votre compte n\'est pas activé.';
 						}
-					}else {
+					}
+					else{
 						$errors[] = 'Le couple identifiant/mot de passe est invalide';
 					}
 				}
@@ -504,8 +511,8 @@ class UserController extends MasterController
 	/**
 	 * Fonction permettant de recevoir un nouveau mot de passe
 	 */
-	public function getNewPassword(){
-
+	public function getNewPassword()
+	{
 		$error = [];
 		$post = [];
 		$success = false;
@@ -559,7 +566,7 @@ class UserController extends MasterController
 							$mail->Password = 'f693c45ea37fa0';                   // SMTP password
 							$mail->SMTPSecure = 'tls';                            // Enable TLS encryption, `ssl` also accepted
 							$mail->Port = 465;                                    // TCP port to connect to
-							
+
 		    	        	$mail->setFrom('limowf3@yopmail.com', 'contact du site'); //expéditeur
 		    	        	$mail->addAddress($post['email'], '');  // Add a recipient// Name is optional
 		    	        	$mail->addReplyTo('info@example.com', 'Information');// si on l'enlève ça renvoie auto à l'expéditeur
@@ -596,84 +603,89 @@ class UserController extends MasterController
 	/**
 	 * Function Permettant de demander un nouveau password
 	 */
-	public function lostPassword(){
-
-		$error = [];
-		$post = [];
-		$showFormPassword = false; // On affiche le 2nd formulaire de mise à jour de notre mdp
-		$showConnectButton = false;
-
-
-		if(isset($_GET['token']) && !empty($_GET['token']) && isset($_GET['email']) && !empty($_GET['email'])){
-
-			// On récupère et nétoie les valeur des GET
-			foreach ($_GET as $key => $value) {
-				$get[$key] = trim(strip_tags($value));
-			}
-
-			$email = $get['email'];
-			$token = $get['token'];
-
-			$tokenModel = new TokensPasswordModel();
-
-			// On vérifie que le token et l'email sont bien dans la base de données
-	  		$tokenExist = $tokenModel->findTokenPassword($email, $token);
-
-	  		if(!empty($tokenExist) && ($tokenExist['date_exp'] > date('Y-m-d H:i:s'))) {
-	  			// Ici le token est valide et n'a pas expiré, on peut donc afficher le formulaire.
-				$showFormPassword = true;
-			}
-			if($tokenExist['date_exp'] < date('Y-m-d H:i:s')){
-				$error[] = 'Le lien n\'est plus valide.' ;
-			}
-		}//fin if post action
-
-		// Traitement du 2nd formulaire concernant la maj du mdp
-		if(isset($_POST['action']) && $_POST['action'] == 'updatePassword') {
-
-	    foreach ($_POST as $key => $value) {
-	        $post[$key] = trim(strip_tags($value));
-		    }
-
-	    if(strlen($post['new_password']) < 8 || strlen($post['new_password']) > 25 ) { // Nbres de caractères modifiables
-				$error[] = 'Votre mot de passe doit contenir entre 8 et 25 caractères';
-			}
-		if($post['new_password'] != $post['new_password_conf']) {
-			$error[] = 'Vos mots de passe doivent correspondre';
+	public function lostPassword()
+	{
+		if(!isset($_GET['email']) && !isset($_GET['token'])){
+			$this->redirectToRoute('default_home');
 		}
-		if(count($error) == 0 ) { // Pas d'erreurs, on continue
-			// Ici, on peut changer le mot de passe
+		else{
+			$error = [];
+			$post = [];
+			$showFormPassword = false; // On affiche le 2nd formulaire de mise à jour de notre mdp
+			$showConnectButton = false;
 
-			// On "Hash" le password
-        	$authModel = new AuthModel();
-			$password = $authModel->hashPassword($post['new_password']);
 
-			// On récupère les info de la table user grace à son adresse mail
-			$usersModel = new UsersModel();
-			$infoUser = $usersModel->getUserByUsernameOrEmail($post['email']);
+			if(isset($_GET['token']) && !empty($_GET['token']) && isset($_GET['email']) && !empty($_GET['email'])){
 
-			$data = [
-				'password' => $password,
-			];
+				// On récupère et nétoie les valeur des GET
+				foreach ($_GET as $key => $value) {
+					$get[$key] = trim(strip_tags($value));
+				}
 
-			echo '<hr>'.$infoUser['id'];
-			if($usersModel->update($data, $infoUser['id'])){
-	        // Suppression du token car le mdp est modifié
-					
-				if($tokenModel->delete($tokenExist['id'])){
-					$showFormPassword = false;
-					$showConnectButton = true;
-            	}
-			}else {
-				$error[] = 'Erreur lors du changement de mot de passe, si le problème persiste veuillez contacter l\'administrateur.';
+				$email = $get['email'];
+				$token = $get['token'];
+
+				$tokenModel = new TokensPasswordModel();
+
+				// On vérifie que le token et l'email sont bien dans la base de données
+		  		$tokenExist = $tokenModel->findTokenPassword($email, $token);
+
+		  		if(!empty($tokenExist) && ($tokenExist['date_exp'] > date('Y-m-d H:i:s'))) {
+		  			// Ici le token est valide et n'a pas expiré, on peut donc afficher le formulaire.
+					$showFormPassword = true;
+				}
+				if($tokenExist['date_exp'] < date('Y-m-d H:i:s')){
+					$error[] = 'Le lien n\'est plus valide.' ;
+				}
+			}//fin if post action
+
+			// Traitement du 2nd formulaire concernant la maj du mdp
+			if(isset($_POST['action']) && $_POST['action'] == 'updatePassword') {
+
+		    foreach ($_POST as $key => $value) {
+		        $post[$key] = trim(strip_tags($value));
+			    }
+
+		    if(strlen($post['new_password']) < 8 || strlen($post['new_password']) > 25 ) { // Nbres de caractères modifiables
+					$error[] = 'Votre mot de passe doit contenir entre 8 et 25 caractères';
+				}
+			if($post['new_password'] != $post['new_password_conf']) {
+				$error[] = 'Vos mots de passe doivent correspondre';
 			}
-		} // fin if count error
-    	else{
-          $error[] = 'Le token et l\'adresse email ne correspondent pas.';
-    	} //fin else
-		}//fin if isset
-		$params = ['error' => $error, 'showFormPassword' => $showFormPassword, 'showConnectButton' => $showConnectButton];
-		$this->show('user/lostPassword', $params);
+			if(count($error) == 0 ) { // Pas d'erreurs, on continue
+				// Ici, on peut changer le mot de passe
+
+				// On "Hash" le password
+	        	$authModel = new AuthModel();
+				$password = $authModel->hashPassword($post['new_password']);
+
+				// On récupère les info de la table user grace à son adresse mail
+				$usersModel = new UsersModel();
+				$infoUser = $usersModel->getUserByUsernameOrEmail($post['email']);
+
+				$data = [
+					'password' => $password,
+				];
+
+				echo '<hr>'.$infoUser['id'];
+				if($usersModel->update($data, $infoUser['id'])){
+		        // Suppression du token car le mdp est modifié
+
+					if($tokenModel->delete($tokenExist['id'])){
+						$showFormPassword = false;
+						$showConnectButton = true;
+	            	}
+				}else {
+					$error[] = 'Erreur lors du changement de mot de passe, si le problème persiste veuillez contacter l\'administrateur.';
+				}
+			} // fin if count error
+	    	else{
+	          $error[] = 'Le token et l\'adresse email ne correspondent pas.';
+	    	} //fin else
+			}//fin if isset
+			$params = ['error' => $error, 'showFormPassword' => $showFormPassword, 'showConnectButton' => $showConnectButton];
+			$this->show('user/lostPassword', $params);
+		}
 	}
 
 	/**
